Move Lists to its proper folder and fixed a warning
